feat(*): enable requiring a deis/charts commit
We already are set up to parse a required commit in any deis repo
provided it is mentioned in the git commit body, e.g. `Requires deis/<repo>#<pr_number>`

(Previously, we only actually performed subsequent logic if it was a Workflow component repo.)

This sets us up so that we can perform logic if the repo is deis/charts.  This logic consists of
determining the appropriate repo clone url alongside the charts commit sha and passing this info
down do the downstream e2e job so that it can set up its local helmc instance appropriately,
before fetching charts.

Closes https://github.com/deis/jenkins-jobs/issues/151