fix(bash/scripts/get_actual_commit.sh): inspect ghprbActualCommit
We'd previously tried not to rely on the environment for getting actual PR commits
(specifically depending on the `ghprbActualCommit` env var provided by the GitHub Pull Request Builder Plugin)
and instead invoked the `git` cli and/or curled the GitHub API directly, as a general practice
of reducing dependencies on plugin code as well as increasing portability between jobs defined in DSL here
and in Jenkinsfile logic, etc.

Blah, blah, blah, anyways, since we _are_ still using the ghprb plugin, and we've seen [issues](https://github.com/deis/jenkins-jobs/issues/227)
with solely relying on the GitHub API, this change brings back logic to defer to the `ghprbActualCommit`
env var, when present, as the PR commit for building artifacts and updating statuses. (We'll still curl the api as a backup if not present.)