ref(*): POC notify Slack via curl
This POC refactors the component build jobs to notify slack via its incoming webhook
api, forgoing the use of the Slack Notifications plugin.  The big win here is the ability
to send a message to a channel other than the global default (which the aforementioned
plugin does not currently allow in a secure manner).